api/requirements.txt
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cliapp==1.20180812.1
click==8.1.5
cloudpickle==2.2.1
cmake==3.26.4
cmdtest==0.32+git
comm==0.1.4
command-not-found==0.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
Cython==3.0.0
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
designer==0.4.2
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
easyocr==1.7.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
fonttools==4.41.1
frozendict==2.3.8
greenlet==2.0.2
Historic-Crypto==0.1.6
html5lib==1.1
httplib2==0.20.2
idna==3.4
imageio==2.31.1
importlib-metadata==4.6.4
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lit==16.0.6
lxml==4.9.3
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==8.10.0
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.1
ninja==1.11.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.5.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post4
pycparser==2.21
pygame==2.5.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.18.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
python-apt==2.4.0+ubuntu1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.1.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
s3transfer==0.6.1
scikit-image==0.21.0
scipy==1.11.1
SecretStorage==3.3.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
spyder-kernels==2.4.4
SQLAlchemy==2.0.19
stack-data==0.6.2
sympy==1.12
systemd-python==234
tifffile==2023.7.10
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
ttystatus==0.38
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.16
virtualenv==20.13.0+ds
wadllib==1.3.6
Wand==0.6.11
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
wurlitzer==3.0.3
yfinance==0.2.28
zipp==1.0.0
